Update versions and fix tests (#1)

+++ src/error.rs
@@ -2,7 +2,7 @@ use thiserror::Error;
 
 pub type FsDkrResult<T> = Result<T, FsDkrError>;
 
-#[derive(Error, Debug)]
+#[derive(Error, Debug, Clone)]
 pub enum FsDkrError {
     #[error("Too many malicious parties detected! Threshold {threshold:?}, Number of Refreshed Messages: {refreshed_keys:?}, Malicious parties detected when trying to refresh: malicious_parties:?")]
     PartiesThresholdViolation {
@@ -30,7 +30,7 @@ pub enum FsDkrError {
     },
 
     #[error("Paillier verification proof failed for party {party_index:?}")]
-    PaillierVerificationError { party_index: usize },
+    PaillierVerificationError { party_index: u16 },
 
     #[error("A new party did not receive a valid index.")]
     NewPartyUnassignedIndexError,
@@ -39,5 +39,5 @@ pub enum FsDkrError {
     BroadcastedPublicKeyError,
 
     #[error("DLog proof failed for party {party_index:?}")]
-    DLogProofValidation { party_index: usize },
+    DLogProofValidation { party_index: u16 },
 }
